game is showing
