+++ Frankie Tsoi/week7/tic-tac-toe-react/src/Game.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import './index.css'; 
+import Board from './Board';
+
+
+class Game extends React.Component {
+  render() {
+    return (
+      <div className="game">
+        <div className="game-board">
+          <Board />
+        </div>
+        <div className="game-info">
+          <div>{/* status */}</div>
+          <ol>{/* TODO */}</ol>
+        </div>
+      </div>
+    ); //return
+  } // render
+}// class Game
+
+export default Game;
