+++ Frankie Tsoi/week7/tic-tac-toe-react/src/Square.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css'; 
 
 
@@ -7,61 +6,21 @@ class Square extends React.Component{
 
   render(){
     return(
-      <button className="square">{/* TODO */}</button>
+
+      <button className="square" onClick={function(){console.log('click');}}> 
+      
+      
+      {this.props.value}</button>
     ); // return
   } // render
 
 }// class Square
 
-class Board extends React.Component{
-  renderSquare(i){
-    return <Square/>;
-  }// render
 
-  render() {
-    const status = 'Next Player: X';
-    return(
-      <div>
-        <div className="status">{status}</div>
-        <div className="status">{status}</div>
-        <div className="board-row">
-          {this.renderSquare(0)}
-          {this.renderSquare(1)}
-          {this.renderSquare(2)}
-        </div>
-        <div className="board-row">
-          {this.renderSquare(3)}
-          {this.renderSquare(4)}
-          {this.renderSquare(5)}
-        </div>
-        <div className="board-row">
-          {this.renderSquare(6)}
-          {this.renderSquare(7)}
-          {this.renderSquare(8)}
-        </div>
-      </div>
-    );//return
-  }//render
 
-}// class Board
 
-class Game extends React.Component{
-  render() {
-    return(
-      <div className="game">
-        <div className="game-board">
-          <Board/>
-        </div>
-        <div className="game-info">
-          <div>{/*status*/}</div>
-          <ol>{/*TODO*/}</ol>
-        </div>
-      </div>
-    );// return
-  }//render
-}// class Game
 
-ReactDOM.render(
-  <Game />,
-  document.getElementById('root')
-)
+
+
+
+  export default Square;
