+++ Frankie Tsoi/week7/tic-tac-toe-react/src/index.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Square from './Square';
+import Game from './Game';
 
 
 ReactDOM.render(
-  
-    <Square />,
- 
+    <Game />,
   document.getElementById('root')
 );
+
